Pad class refractor init
